Configuration:  - Refactor class hierarchy to ease implementation of configuration repositories  - Expose creation of default repository as public, to enable its use by client projects  - Add some javadoc
