+++ commons/src/main/java/com/powsybl/commons/config/StackedModuleConfig.java
@@ -16,7 +16,7 @@ import java.util.function.Supplier;
  *
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class StackedModuleConfig extends AbstractMapModule {
+public class StackedModuleConfig extends AbstractModuleConfig {
 
     private final ModuleConfig config1;
